Create muslimwebdev
Hi there! I'm Muslim Bashir, a passionate web developer with a knack for creating dynamic and responsive web applications. With a background in both front-end and back-end development, I enjoy bringing ideas to life in the browser. I'm constantly learning and evolving, embracing new technologies and best practices in the web development world.

Here are some of the technologies and tools I work with:

- **Front-end:** HTML, CSS, JavaScript, React, Vue.js, Bootstrap.
- **Back-end:** Node.js,PHP.
- **Databases:** MySQL
- **Version Control:** Git, GitHub